+++ egui_demo_lib/src/apps/demo/mod.rs
@@ -7,6 +7,7 @@
 mod app;
 pub mod code_editor;
 pub mod code_example;
+pub mod context_menu;
 pub mod dancing_strings;
 pub mod demo_app_windows;
 pub mod drag_and_drop;