+++ CHANGELOG.md
@@ -7,6 +7,16 @@ NOTE: [`epaint`](epaint/CHANGELOG.md), [`eframe`](eframe/CHANGELOG.md), [`egui_w
 
 ## Unreleased
 
+### Added ⭐
+* Add context menus: See `Ui::menu_button` and `Response::context_menu` ([#543](https://github.com/emilk/egui/pull/543)).
+* You can now read and write the cursor of a `TextEdit` ([#848](https://github.com/emilk/egui/pull/848)).
+
+### Changed 🔧
+* Unifiy the four `Memory` data buckets (`data`, `data_temp`, `id_data` and `id_data_temp`) into a single `Memory::data`, with a new interface ([#836](https://github.com/emilk/egui/pull/836)).
+
+### Contributors 🙏
+* [mankinskin](https://github.com/mankinskin) ([#543](https://github.com/emilk/egui/pull/543))
+
 
 ## 0.15.0 - 2021-10-24 - Syntax highlighting and hscroll